+++ src/main/kotlin/csense/idea/kotlin/assistance/inspections/SimpleInfiniteRecursionInspection.kt
@@ -1,18 +1,12 @@
 package csense.idea.kotlin.assistance.inspections
 
-import com.intellij.codeHighlighting.HighlightDisplayLevel
-import com.intellij.codeInspection.ProblemsHolder
-import com.intellij.codeInspection.SuppressIntentionAction
-import com.intellij.psi.PsiElement
-import csense.idea.base.bll.psi.findParentOfType
-import csense.idea.kotlin.assistance.Constants
-import csense.idea.kotlin.assistance.suppression.KtExpressionSuppression
-import org.jetbrains.kotlin.idea.inspections.AbstractKotlinInspection
-import org.jetbrains.kotlin.nj2k.postProcessing.resolve
-import org.jetbrains.kotlin.psi.KtCallExpression
-import org.jetbrains.kotlin.psi.KtFunction
-import org.jetbrains.kotlin.psi.KtVisitorVoid
-import org.jetbrains.kotlin.psi.callExpressionVisitor
+import com.intellij.codeHighlighting.*
+import com.intellij.codeInspection.*
+import com.intellij.psi.*
+import csense.idea.kotlin.assistance.*
+import csense.idea.kotlin.assistance.suppression.*
+import org.jetbrains.kotlin.idea.inspections.*
+import org.jetbrains.kotlin.psi.*
 
 //eg fun x()= x()
 //or no throwing of exceptions and no returns except for calling the function again...
@@ -55,8 +49,10 @@ class SimpleInfiniteRecursionInspection : AbstractKotlinInspection() {
         return HighlightDisplayLevel.WARNING
     }
 
-    override fun buildVisitor(holder: ProblemsHolder,
-                              isOnTheFly: Boolean): KtVisitorVoid {
+    override fun buildVisitor(
+            holder: ProblemsHolder,
+            isOnTheFly: Boolean
+    ): KtVisitorVoid {
         return callExpressionVisitor { exp: KtCallExpression ->
 //            if (exp.resolve() == exp.findParentOfType<KtFunction>()) {
 //                holder.registerProblem(exp, "todo must see more.")